requirements.txt
# Enhanced Video Dubbing Automation - Requirements
# Core ML and Audio Processing
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
datasets>=2.10.0

# Speech Processing
openai-whisper>=20231117
speechbrain>=0.5.15
librosa>=0.10.0
soundfile>=0.12.0
noisereduce>=3.0.0

# Video Processing
moviepy>=1.0.3
opencv-python>=4.8.0
ffmpeg-python>=0.2.0

# Translation and NLP
sentencepiece>=0.1.99
sacremoses>=0.0.53
tokenizers>=0.13.0

# Synchronization
dtw-python>=1.3.0
scipy>=1.10.0
numpy>=1.24.0

# Utilities and System
psutil>=5.9.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0

# Git and Repository Management
gitpython>=3.1.0

# Optional: Jupyter and Visualization (for local development)
jupyter>=1.0.0
ipywidgets>=8.0.0
plotly>=5.15.0

# Note: Some packages may be pre-installed on Kaggle
# Install with: pip install -r requirements.txt
# For Kaggle: pip install --user -r requirements.txt
